+++ dzVents/runtime/Time.lua
@@ -106,7 +106,7 @@ function getYearBeginDayOfWeek(tm)
 	return yearBeginDayOfWeek
 end
 
--- tm: date (as retruned fro os.time)
+-- tm: date (as returned fro os.time)
 -- returns basic correction to be add for counting number of week
 -- weekNum = math.floor((dayOfYear + returnedNumber) / 7) + 1
 -- (does not consider correction at begin and end of year)
@@ -237,7 +237,8 @@ local function Time(sDate, isUTC, _testMS)
 	end
 
 	self.rawDate = self.year .. '-' .. string.format("%02d", self.month) .. '-' .. string.format("%02d", self.day)
-	self.rawTime = string.format("%02d", self.hour) .. ':' .. string.format("%02d", self.min) .. ':' .. string.format("%02d", self.sec)
+	self.time = string.format("%02d", self.hour) .. ':' .. string.format("%02d", self.min)
+    self.rawTime =  self.time .. ':' .. string.format("%02d", self.sec)
 	self.rawDateTime = self.rawDate .. ' ' .. self.rawTime
 	self.milliSeconds = ms
 	self.milliseconds = ms
@@ -262,7 +263,8 @@ local function Time(sDate, isUTC, _testMS)
 	self['minutes'] = self.min
 	self['seconds'] = self.sec
 
-	self['secondsSinceMidnight'] = self.hour * 3600 + self.min * 60 + self.sec
+	self['minutesSinceMidnight'] = self.hour * 60 + self.min
+	self['secondsSinceMidnight'] = self.minutesSinceMidnight * 60 + self.sec
 	self['utils'] = utils
 	self['isUTC'] = isUTC
 	self['dDate'] = dDate
@@ -296,6 +298,31 @@ local function Time(sDate, isUTC, _testMS)
 		else
 			utils.log('Invalid time format passed to diff. Should a Time object', utils.LOG_ERROR)
 		end
+	end
+
+	function self.addSeconds(seconds, factor)
+		if type(seconds) ~= 'number' then 
+			self.utils.log(tostring(seconds) .. ' is not a valid parameter to this function. Please change to use a number value!', utils.LOG_ERROR)
+		else
+			factor = factor or 1
+			return Time( os.date("%Y-%m-%d %X", self.dDate +  factor * math.floor(seconds) )) 
+		end
+	end
+
+	function self.addDays(days)
+		return self.addSeconds(days, 24 * 3600)
+	end
+
+	function self.addHours(hours)
+		return self.addSeconds(hours, 3600)
+	end
+
+	function self.addMinutes(minutes)
+		return self.addSeconds(minutes, 60)
+	end
+
+	function self.makeTime(sDate, isUTC)
+		return Time(sDate, isUTC)
 	end
 
 	-- return ISO format
@@ -378,8 +405,8 @@ local function Time(sDate, isUTC, _testMS)
 		local testH = self.hour
 		local testM = self.min
 
-		if (stopH < startH) then -- add 24 hours if endhours < startHours
-			local stopHOrg = stopH
+		if (( stopH * 24 + stopM ) < ( startH * 24 + startM ) ) then -- add 24 hours if endTime < startTime
+            		local stopHOrg = stopH
 			stopH = stopH + 24
 			if (testH <= stopHOrg) then -- if endhours has increased the currenthour should also increase
 				testH = testH + 24
@@ -835,7 +862,7 @@ local function Time(sDate, isUTC, _testMS)
 
 	-- returns true if self.time is in time range: at hh:mm-hh:mm
 	function self.ruleMatchesTimeRange(rule)
-		local fromH, fromM, toH, toM = string.match(rule, 'at% ([0-9%*]+):([0-9%*]+)-([0-9%*]+):([0-9%*]+)')
+		local fromH, fromM, toH, toM =  string.match(rule, 'at% ([0-9%*]+):([0-9%*]+)-([0-9%*]+):([0-9%*]+)') 
 
 		if (fromH ~= nil) then
 			-- all will be nil if fromH is nil
@@ -962,8 +989,17 @@ local function Time(sDate, isUTC, _testMS)
 
 		return timeIsInRange(fromHH, fromMM, toHH, toMM)
 	end
-
-	-- returns true if self.time matches the rule
+    
+    -- remove seconds from timeStrings 'at 12:23:56-23:45:00 ==>> 'at 12:23-23:45' to allow use of rawTime in matchesRule
+	local function sanitize(rule)
+        if not rule:match("(%w+%:%w+:%w+)") then return rule end
+        for strippedTime in rule:gmatch("(%w+%:%w+)") do
+            rule = rule:gsub(rule:match("(%w+%:%w+:%w+)"),rule:match(strippedTime))
+        end
+        return rule
+    end
+    
+    -- returns true if self.time matches the rule
 	function self.matchesRule(rule)
 		if (string.len(rule == nil and "" or rule) == 0) then
 			return false
@@ -1136,14 +1172,15 @@ local function Time(sDate, isUTC, _testMS)
 		end
 		updateTotal(res)
 
-		res = self.ruleMatchesTime(rule) -- moment / range
+        rule = sanitize(rule)    
+        res = self.ruleMatchesTime(rule) -- moment / range
 		if (res == false) then
 			-- rule had at hh:mm part but didn't match (or was invalid)
 			return false
 		end
 		updateTotal(res)
 
-		res = self.ruleMatchesTimeRange(rule) -- range
+		res = self.ruleMatchesTimeRange(sanitize(rule)) -- range
 		if (res == false) then
 			-- rule had at hh:mm-hh:mm but time is not in that range now
 			return false