+++ dzVents/runtime/tests/scripts/script_wildcard3.lua
@@ -1,0 +1,9 @@
+return {
+	active = true,
+	on = {
+		devices = {'*(test)*'}
+	},
+	execute = function(domoticz, device)
+		return 'script_wildcard1: ' .. domoticz.name .. ' ' .. device.name
+	end
+}