+++ hardware/EvohomeRadio.cpp
@@ -40,6 +40,8 @@ enum evoCommands
 	cmdDHWTemp = 0x1260,
 	cmdControllerMode = 0x2E04,
 	cmdControllerHeatDemand = 0x0008,//Heat demand sent by the controller for CH / DHW / Boiler  (F9/FA/FC)
+	cmdOpenThermBridge = 0x3220,//OT Bridge Status messages
+        cmdOpenThermSetpoint = 0x22D9,//OT Bridge Control Setpoint
 	cmdActuatorState = 0x3EF0,
 	cmdActuatorCheck = 0x3B00,
 	cmdBinding = 0x1FC9,
@@ -93,6 +95,8 @@ CEvohomeRadio::CEvohomeRadio(const int ID, const std::string &UserContID)
 	RegisterDecoder(cmdSysInfo, boost::bind(&CEvohomeRadio::DecodeSysInfo, this, _1));
 	RegisterDecoder(cmdZoneName, boost::bind(&CEvohomeRadio::DecodeZoneName, this, _1));
 	RegisterDecoder(cmdZoneHeatDemand, boost::bind(&CEvohomeRadio::DecodeHeatDemand, this, _1));
+	RegisterDecoder(cmdOpenThermBridge, boost::bind(&CEvohomeRadio::DecodeOpenThermBridge, this, _1));
+        RegisterDecoder(cmdOpenThermSetpoint, boost::bind(&CEvohomeRadio::DecodeOpenThermSetpoint, this, _1));
 	RegisterDecoder(cmdZoneInfo, boost::bind(&CEvohomeRadio::DecodeZoneInfo, this, _1));
 	RegisterDecoder(cmdControllerHeatDemand, boost::bind(&CEvohomeRadio::DecodeHeatDemand, this, _1));
 	RegisterDecoder(cmdBinding, boost::bind(&CEvohomeRadio::DecodeBinding, this, _1));
@@ -449,8 +453,10 @@ void CEvohomeRadio::SendExternalSensor()
 	{
 		std::vector<std::string> strarray;
 		StringSplit(result[0][0], ";", strarray);
-		if (!strarray.empty())
-			dbTemp = atof(strarray[0].c_str());
+		if (!strarray.empty()) {
+		    dbTemp = atof(strarray[0].c_str());
+            AddSendQueue(CEvohomeMsg(CEvohomeMsg::pktinf, 0, GetGatewayID(), cmdExternalSensor).Add((uint8_t)2).Add(static_cast<int16_t>(dbTemp*100.0)).Add((uint8_t)1));
+        }
 		else
 			return;
 	}
@@ -459,12 +465,12 @@ void CEvohomeRadio::SendExternalSensor()
 
 	//FIXME no light level data available UV from WU is only thing vaguely close (on dev system) without a real sensor
 	result = m_sql.safe_query("SELECT sValue FROM DeviceStatus WHERE (Type==%d)", (int)pTypeUV);
-	if (!result.empty())
-		dbUV = atof(result[0][0].c_str());
+	if (!result.empty()) {
+	    dbUV = atof(result[0][0].c_str());
+        AddSendQueue(CEvohomeMsg(CEvohomeMsg::pktinf, 0, GetGatewayID(), cmdExternalSensor).Add((uint8_t)0).Add(static_cast<uint16_t>(dbUV * 39)).Add((uint8_t)2));
+    }
 	else
 		return;
-
-	AddSendQueue(CEvohomeMsg(CEvohomeMsg::pktinf, 0, GetGatewayID(), cmdExternalSensor).Add((uint8_t)0).Add(static_cast<uint16_t>(dbUV * 39)).Add((uint8_t)2).Add((uint8_t)2).Add(static_cast<int16_t>(dbTemp*100.0)).Add((uint8_t)1));
 }
 
 
@@ -1445,6 +1451,49 @@ bool CEvohomeRadio::DecodeHeatDemand(CEvohomeMsg &msg)
 	return true;
 }
 
+void CEvohomeRadio::UpdateSwitch(const unsigned char Idx, const bool bOn, const std::string &defaultname)
+{
+	char szIdx[10];
+	sprintf(szIdx,"%X%02X%02X%02X",0,0,0,Idx);
+	std::vector<std::vector<std::string> > result;
+	result = m_sql.safe_query("SELECT Name,nValue,sValue FROM DeviceStatus WHERE (HardwareID==%d) AND (DeviceID=='%q')", m_HwdID, szIdx);
+	if (!result.empty())
+	{
+		//check if we have a change, if not do not update it
+		int nvalue=atoi(result[0][1].c_str());
+		if ((!bOn)&&(nvalue==0))
+			return;
+		if ((bOn&&(nvalue!=0)))
+			return;
+	}
+
+	//Send as Lighting 2
+	tRBUF lcmd;
+	memset(&lcmd,0,sizeof(RBUF));
+	lcmd.LIGHTING2.packetlength=sizeof(lcmd.LIGHTING2)-1;
+	lcmd.LIGHTING2.packettype=pTypeLighting2;
+	lcmd.LIGHTING2.subtype=sTypeAC;
+	lcmd.LIGHTING2.id1=0;
+	lcmd.LIGHTING2.id2=0;
+	lcmd.LIGHTING2.id3=0;
+	lcmd.LIGHTING2.id4=Idx;
+	lcmd.LIGHTING2.unitcode=1;
+	int level=15;
+	if (!bOn)
+	{
+		level=0;
+		lcmd.LIGHTING2.cmnd=light2_sOff;
+	}
+	else
+	{
+		level=15;
+		lcmd.LIGHTING2.cmnd=light2_sOn;
+	}
+	lcmd.LIGHTING2.level= (uint8_t)level;
+	lcmd.LIGHTING2.filler=0;
+	lcmd.LIGHTING2.rssi=12;
+	sDecodeRXMessage(this, (const unsigned char *)&lcmd.LIGHTING2, defaultname.c_str(), 255);
+}
 
 bool CEvohomeRadio::DecodeActuatorCheck(CEvohomeMsg &msg)
 {
@@ -1468,6 +1517,31 @@ bool CEvohomeRadio::DecodeActuatorCheck(CEvohomeMsg &msg)
 bool CEvohomeRadio::DecodeActuatorState(CEvohomeMsg &msg)
 {
 	char tag[] = "ACTUATOR_STATE";
+        // If there is an OT Bridge the controller RQs the OT Bridge every minute with a payload size of 1
+        if (msg.payloadsize == 1) {
+        //      A payload of size 1 is normally just 0x00
+                return true;
+        }
+        // The OT Bridge responds to the RQ with a RP with payload size of 6
+        if (msg.payloadsize == 6) {
+		// msg.payload[1] is the Relative modulation level
+                bool bExists = CheckPercentageSensorExists(17, 1);
+		if ((msg.payload[1] != 0) || (bExists))
+                {
+                        SendPercentageSensor(17, 1, 255,  static_cast<float>(msg.payload[1]), "Relative modulation level");
+                }
+ 		// msg.payload[3] is the Flame Status
+		bool bFlameOn { false };
+		if (msg.payload[3] == 0x0A) { 
+			bFlameOn = true;
+		} 
+		// Record the Flame status in the same way that the OTGW hardware device currently does
+		UpdateSwitch(113,bFlameOn,"FlameOn");
+
+                Log(true, LOG_STATUS, "evohome: %s: OT Bridge full payload %02X%02X%02X%02X%02X%02X, packet size: %d", tag, msg.payload[0], msg.payload[1], msg.payload[2], msg.payload[3], msg.payload[4], msg.payload[5], msg.payloadsize);
+                return true;
+        }
+ 	// All other relays should have a payload size of 3	
 	if (msg.payloadsize != 3) {
 		Log(false, LOG_ERROR, "evohome: %s: Error decoding command, unknown packet size: %d", tag, msg.payloadsize);
 		return false;
@@ -1481,6 +1555,127 @@ bool CEvohomeRadio::DecodeActuatorState(CEvohomeMsg &msg)
 	return true;
 }
 
+bool CEvohomeRadio::DecodeOpenThermBridge(CEvohomeMsg &msg)
+{
+        char tag[] = "OPENTHERM_BRIDGE";
+
+	// Only look for responses from the OT Bridge and Filter out messages from other controllers 
+	if (msg.GetID(1) != GetControllerID()) 
+		return true;
+
+	// All OT messages should have a payload size of 5	
+	if (msg.payloadsize != 5) {
+		Log(false, LOG_ERROR, "evohome: %s: Error decoding command, unknown packet size: %d", tag, msg.payloadsize);
+		return false;
+	}
+	// The OT command response is in byte 4 and 5
+	int nOTResponse = msg.payload[3] << 8 | msg.payload[4];
+	float fOTResponse = static_cast<float>(nOTResponse) / 256.0f;
+
+	// The OT commands are as per the OT Specification
+	// 05 (ID.05) = Fault Code
+	if (msg.payload[2] == 0x05) {
+		if (msg.payload[3] != 0)
+		{
+			SendCustomSensor(0, 5, 255, static_cast<float>(msg.payload[3]), "Application fault flags", ""); 
+		}
+		// Note : Vaillant VR33 returns back 255 as normal 
+                if ((msg.payload[4] != 0) && (msg.payload[4] != 0xFF))
+                {
+                        SendCustomSensor(1, 5, 255, static_cast<float>(msg.payload[4]), "OEM fault code", "");
+                }
+		Log(true, LOG_STATUS, "evohome: %s: Application-specific flags = %02X %d", tag, msg.payload[3],  msg.payload[4]);
+		return true;
+	}
+	// 11 (ID.17) = Relative modulation level
+	if (msg.payload[2] == 0x11) {
+		bool bExists = CheckPercentageSensorExists(17, 1);
+		if ((fOTResponse != 0) || (bExists))
+		{
+			SendPercentageSensor(17, 1, 255, fOTResponse, "Relative modulation level");
+		}
+		Log(true, LOG_STATUS, "evohome: %s: Relative modulation level = %.2f %%", tag, fOTResponse);
+		return true;
+	}
+	// 12 (ID.18) = CH water pressure
+	if (msg.payload[2] == 0x12) {
+                if (fOTResponse != 0)
+                {
+			SendPressureSensor(0, 18, 255, fOTResponse, "CH Water Pressure");
+		}
+		Log(true, LOG_STATUS, "evohome: %s: CH water pressure = %.2f bar", tag, fOTResponse);
+		return true;
+	}
+        // 13 (ID.19) = DHW flow rate
+        if (msg.payload[2] == 0x13) {
+                if (fOTResponse != 0)
+                {
+			SendWaterflowSensor(0, 19, 255, fOTResponse, "DHW flow rate");
+		}
+        	Log(true, LOG_STATUS, "evohome: %s: DHW flow rate = %.2f l/min", tag, fOTResponse);
+        	return true;
+        }
+	// 19 (ID.25) = Boiler Water Temperature
+	if (msg.payload[2] == 0x19) {
+                if (fOTResponse != 0)
+                {
+ 	        	SendTempSensor(25, 255, fOTResponse, "Boiler Water Temperature");
+		}
+ 		Log(true, LOG_STATUS, "evohome: %s: Boiler Water Temperature = %.2f C", tag, fOTResponse);
+		return true;
+	}
+	// 1A (ID.26) = DHW Temperature
+	if (msg.payload[2] == 0x1a) {
+		if (fOTResponse != 0)
+                {
+ 			SendTempSensor(26, 255, fOTResponse, "DHW Temperature");
+		}
+		Log(true, LOG_STATUS, "evohome: %s: DHW Temperature = %.2f C", tag, fOTResponse);
+		return true;
+	}
+	// 1C (ID.28) = Return Water Temperature
+	if (msg.payload[2] == 0x1c) {
+                if (fOTResponse != 0)
+                {
+  			SendTempSensor(28, 255, fOTResponse, "Return Water Temperature");
+		}
+ 		Log(true, LOG_STATUS, "evohome: %s: Return Water Temperature = %.2f C", tag, fOTResponse);
+		return true;
+	}
+	// 73 (ID.115) = OEM diagnostic code
+	if (msg.payload[2] == 0x73) {
+                if (nOTResponse != 0)
+                {
+                        SendCustomSensor(0, 115, 255, static_cast<float>(nOTResponse), "OEM diagnostic code", "");
+		} 
+		Log(true, LOG_STATUS, "evohome: %s: OEM diagnostic code = %d", tag, nOTResponse);
+		return true;
+	}
+	return true;
+}
+
+bool CEvohomeRadio::DecodeOpenThermSetpoint(CEvohomeMsg &msg)
+{
+        char tag[] = "OPENTHERM_SETPOINT";
+
+        // Only look for responses from the OT Bridge and Filter out messages from other controllers
+        if (msg.GetID(1) != GetControllerID())
+                return true;
+
+        // All OT messages should have a payload size of 3
+        if (msg.payloadsize != 3) {
+                Log(false, LOG_ERROR, "evohome: %s: Error decoding command, unknown packet size: %d", tag, msg.payloadsize);
+                return false;
+        }
+
+        // The OT Control Setpoint is in byte 2 and 3
+        float fOTSetpoint = static_cast<float>(msg.payload[1] << 8 | msg.payload[2]) / 100.0f;
+
+	SendTempSensor(1, 255, fOTSetpoint, "Control Setpoint");
+        Log(true, LOG_STATUS, "evohome: %s: Boiler Water Temperature = %.2f C", tag, fOTSetpoint);
+
+        return true;
+}
 
 bool CEvohomeRadio::DecodeExternalSensor(CEvohomeMsg &msg)
 {